requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
python-dotenv==0.21.1
gunicorn==21.2.0
requests==2.31.0
python-chess==1.2.0
pytest==7.4.0
pytest-cov==4.1.0
requests-mock==1.11.0
